Create Main Window
Basic window, not fully complete yet. Still waiting on list of basic ingredients and such, but the main code that will acquire the ingredients and time given is there. Still need to add a text field that stores the microwave option, along with getting all of the boxes to disappear once submit has been pressed. After that, a button that brings you back to the start again.