Update Main
There is now a menu bar at the top of the program. Only tab located is called "Options", and subtabs are "Search" and "Favorites". When the user selects "Search", it takes them back to the search menu to change their search choices. When user selects "Favorites", it brings them to a separate JPane that will display favorited recipes.